+++ Rubeus/Domain/Info.cs
@@ -12,7 +12,7 @@ namespace Rubeus.Domain
             Console.WriteLine("  |  __  /| | | |  _ \\| ___ | | | |/___)");
             Console.WriteLine("  | |  \\ \\| |_| | |_) ) ____| |_| |___ |");
             Console.WriteLine("  |_|   |_|____/|____/|_____)____/(___/\r\n");
-            Console.WriteLine("  v1.3.4\r\n");
+            Console.WriteLine("  v1.3.5\r\n");
         }
 
         public static void ShowUsage()
@@ -83,6 +83,12 @@ Roasting:
 
     Perform Kerberoasting with alternate credentials:
         Rubeus.exe kerberoast /creduser:DOMAIN.FQDN\USER /credpassword:PASSWORD [/spn:""blah/blah""] [/user:USER] [/domain:DOMAIN] [/dc:DOMAIN_CONTROLLER] [/ou:""OU=,...""]
+
+    Perform Kerberoasting with an existing TGT:
+        Rubeus.exe kerberoast /spn:""blah/blah"" </ticket:BASE64 | /ticket:FILE.KIRBI>
+
+    Perform Kerberoasting using the tgtdeleg ticket to requestion service tickets:
+        Rubeus.exe kerberoast /spn:""blah/blah"" /usetgtdeleg
 
     Perform AS-REP ""roasting"" for any users without preauth:
         Rubeus.exe asreproast [/user:USER] [/domain:DOMAIN] [/dc:DOMAIN_CONTROLLER] [/ou:""OU=,...""]